requirements.txt
fastapi==0.105.0 
uvicorn==0.24.0
psycopg2==2.9.9
geojson-pydantic==1.0.1


geojson==3.1.0

## Testing 
pytest==7.4.3
httpx==0.26.0 

## Used for new relic monitoring
newrelic==7.2.4.171
sentry-sdk==1.5.12

## Third party
area==1.1.1
orjson==3.9.10
boto3==1.24.38
fastapi-versioning==0.10.0
redis==5.0.3
celery==5.3.6
slowapi==0.1.8
osm-login-python==1.0.2
humanize==4.9.0
python-slugify==8.0.1
geomet==1.1.0

## documentation
# mkdocs-material==8.5.11
# mkdocs-jupyter==0.22.0
# neoteroi-mkdocs==0.1.2
# pdocs==1.0.1

## flower
flower==2.0.1

##duckdb
duckdb==0.9.2

##hdx
hdx-python-api==6.2.2

## only needed if postgres is used as celery backend 
SQLAlchemy==2.0.25

##sozipfile
sozipfile==0.3.2
## zip memory optimization
zipfly==6.0.5
psutil==5.9.8 


## logging 
tqdm==4.66.2


## 
pydantic-settings==2.2.1